pyproject.toml
[tool.poetry]
name = "pycryptopals"
version = "0.1.0"
description = "Cryptopals in Python"
authors = ["Ricardo Carvalho <rat.carvalho@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pycryptodome = "^3.9.9"

[tool.poetry.dev-dependencies]
pytest = "^6.2.2"
autopep8 = "^1.5.5"
mypy = "^0.812"
black = "^20.8b1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
